aux_hint: Change the time to fetch aux hint to the hook "pre_post_fetch".
The policy that makes client to maximize is applied in "post_fetch",
and aux_hint of "wm.policy.win.user.geometry" was in "pre_new_client" that
is called after "post_fetch".
So, it is better to fetch hints in "pre_post_fetch" that is called
before evaluation of client.

Change-Id: I2a711e313f89b79e3c3c5964cf96ee50d4e7124f