quickpanel: do not hide the handler objects by rotation.
if we hide the handler objects by rotation, the objects can be hidden
unintentionally, because there is no cancel event of rotation of client for now.
so, I just don't hide the handler objects by rotation,
I think there is no reason to do.
and I leave a log for rotation cancel of client.

Change-Id: Iedb1e0f7a76b1a62a30c4e2aab34f8a17374e1f3